+++ sdk/dart/lib/models/access_token.g.dart
@@ -10,7 +10,7 @@ AccessTokenJar _$AccessTokenJarFromJson(Map<String, dynamic> json) {
   return AccessTokenJar(
     json['access_token'] as String,
     json['token_type'] as String,
-    AccessTokenJar._setTokenExpiration(json['expires_in'] as int),
+    json["expires_in"] as int,
     json['refresh_token'] as String,
     json['scope'] as String,
   );
@@ -20,7 +20,7 @@ Map<String, dynamic> _$AccessTokenJarToJson(AccessTokenJar instance) =>
     <String, dynamic>{
       'access_token': instance.accessToken,
       'token_type': instance.type,
-      'expires_in': instance.expires?.toIso8601String(),
+      'expires_in': instance.expires,
       'refresh_token': instance.refreshToken,
       'scope': instance.scope,
     };